Prepare for unit testing
