+++ operator/src/main/java/oracle/kubernetes/operator/calls/CallResponse.java
@@ -1,0 +1,23 @@
+// Copyright 2018, Oracle Corporation and/or its affiliates.  All rights reserved.
+// Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
+
+package oracle.kubernetes.operator.calls;
+
+import io.kubernetes.client.ApiException;
+
+import java.util.List;
+import java.util.Map;
+
+public final class CallResponse<T> {
+  public final T result;
+  public final ApiException e;
+  public final int statusCode;
+  public final Map<String, List<String>> responseHeaders;
+
+  CallResponse(T result, ApiException e, int statusCode, Map<String, List<String>> responseHeaders) {
+    this.result = result;
+    this.e = e;
+    this.statusCode = statusCode;
+    this.responseHeaders = responseHeaders;
+  }
+}
